+++ src/main/java/com/birca/bircabackend/command/cafe/infrastructure/nationaltaxservice/NationalTaxServiceVerificationApi.java
@@ -1,7 +1,5 @@
-package com.birca.bircabackend.command.cafe.infrastructure;
+package com.birca.bircabackend.command.cafe.infrastructure.nationaltaxservice;
 
-import com.birca.bircabackend.command.cafe.dto.BusinessLicenseStatusRequest;
-import com.birca.bircabackend.command.cafe.dto.BusinessLicenseStatusResponse;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -11,6 +9,7 @@ import static org.springframework.util.MimeTypeUtils.APPLICATION_JSON_VALUE;
 
 @FeignClient(name = "nationalTaxServiceVerification", url = "${service-key.key}")
 public interface NationalTaxServiceVerificationApi {
+
     @PostMapping(consumes = APPLICATION_JSON_VALUE)
     ResponseEntity<BusinessLicenseStatusResponse> verifyBusinessLicenseStatus(
             @RequestBody BusinessLicenseStatusRequest request);
